requirements.txt
discord.py>=2.3.0
flask>=2.3.2
